修正记录数据库创建版本功能代码报错的问题
