+++ Core/OTO/Models/index.js
@@ -3,8 +3,6 @@ const path = require("path");
 const Sequelize = require("sequelize");
 const EventManager = require("../../EventManager");
 const Relational = require("./Relational");
-const SystemConfigService = require('../../services/SystemConfig');
-const packageJson = require('../../../package.json');
 
 
 let PO_MODELS = null;//PO对象
@@ -87,7 +85,6 @@ function AutoInit(sqlConnect) {
         console.log("正在初始化数据库......")
         sqlConnect.queryInterface.sequelize.query("PRAGMA foreign_keys = ON");
         sqlConnect.sync(/*{ alter: true }*/).then(result => {
-            initializeDatabase();
             em.emit("DB.Models.Init", sqlConnect.options.storage, result);
         }).catch(err => {
             em.emit("Debug.Log", "数据库同步失败！", "DATABASE", err);
@@ -97,27 +94,6 @@ function AutoInit(sqlConnect) {
     });
 
 }
-
-/**
- * 数据库初始化-初始数据初始化
- */
-async function initializeDatabase() {
-    // 记录数据库初始化时使用的项目版本-便于跟踪后续升级
-    const dbVersion = await SystemConfigService.getConfig(
-        SystemConfigService.Group.DATABASE_VERSION,
-        'create_version'
-    );
-    if (!dbVersion) {
-        await SystemConfigService.setConfig(
-            SystemConfigService.Group.DATABASE_VERSION,
-            'create_version',
-            packageJson.version
-        );
-    }
-}
-
-
-
 
 module.exports = Models;
